requirements.txt
audioop-lts
openai
py-cord
python-dotenv